revert IdentityServicesPropertySourceFactory tests to test in isolation
