+++ spring-security/src/test/java/com/sap/cloud/security/spring/config/IdentityServicesPropertySourceFactoryBrokerNoHoleTest.java
@@ -5,26 +5,18 @@
  */
 package com.sap.cloud.security.spring.config;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.junit.Assert.assertTrue;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.io.IOException;
-import java.util.List;
-
-import com.sap.cloud.environment.servicebinding.SapVcapServicesServiceBindingAccessor;
-import com.sap.cloud.security.config.OAuth2ServiceConfiguration;
-import com.sap.cloud.security.config.Service;
-import com.sap.cloud.security.config.ServiceBindingEnvironment;
 import com.sap.cloud.security.config.ServiceConstants;
-import org.apache.commons.io.IOUtils;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
+
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * Tests that the {@link IdentityServicesPropertySourceFactory} puts 2 XSUAA service instances with plan 'application' into the Spring properties without creating a hole at index 1.
@@ -40,33 +32,22 @@ class IdentityServicesPropertySourceFactoryBrokerNoHoleTest {
 	@Autowired
 	BrokerHoleTestConfigurationFromFile configuration;
 
-	static String serviceBindingJson;
-
-	@BeforeAll
-	static void setup() throws IOException {
-		serviceBindingJson = IOUtils.resourceToString("/xsuaaBindingsTwoApplicationsNoBroker.json", UTF_8);
-	}
-
 	@Test
 	void testInjectedPropertyValues() {
-		ServiceBindingEnvironment env = new ServiceBindingEnvironment(new SapVcapServicesServiceBindingAccessor(any -> serviceBindingJson));
-
 		/* Index 0 */
-		OAuth2ServiceConfiguration xsuaaConfig = env.getXsuaaConfiguration();
-		assertEquals(xsuaaConfig.getClientId(), configuration.xsuaaClientId0);
-		assertEquals(xsuaaConfig.getClientSecret(), configuration.xsuaaClientSecret0);
-		assertEquals(xsuaaConfig.getProperty(ServiceConstants.URL), configuration.xsuaaUrl0);
-		assertEquals(xsuaaConfig.getProperty(ServiceConstants.XSUAA.UAA_DOMAIN), configuration.xsuaaDomain0);
-		assertEquals(xsuaaConfig.getProperty(ServiceConstants.XSUAA.APP_ID), configuration.xsuaaAppName0);
-		assertEquals(ServiceConstants.Plan.APPLICATION, ServiceConstants.Plan.from(configuration.xsuaaPlan0));
+		assertEquals("client-id2", configuration.xsuaaClientId0);
+		assertEquals("client-secret2", configuration.xsuaaClientSecret0);
+		assertEquals("http://domain.xsuaadomain", configuration.xsuaaUrl0);
+		assertEquals("xsuaadomain", configuration.xsuaaDomain0);
+		assertEquals("xsappname2", configuration.xsuaaAppName0);
+		assertEquals("application", configuration.xsuaaPlan0.toLowerCase());
 		assertEquals("", configuration.unknown0);
 
 		/* Index 1 */
-		OAuth2ServiceConfiguration otherXsuaaConfig = env.getServiceConfigurationsAsList().get(Service.XSUAA).stream().filter(c -> c != xsuaaConfig).findFirst().get();
-		assertEquals(otherXsuaaConfig.getClientId(), configuration.xsuaaClientId1);
-		assertEquals(otherXsuaaConfig.getClientSecret(), configuration.xsuaaClientSecret1);
-		assertEquals(otherXsuaaConfig.getProperty(ServiceConstants.XSUAA.APP_ID), configuration.xsuaaAppName1);
-		assertEquals(ServiceConstants.Plan.APPLICATION, ServiceConstants.Plan.from(configuration.xsuaaPlan1));
+		assertEquals("client-id", configuration.xsuaaClientId1);
+		assertEquals("client-secret", configuration.xsuaaClientSecret1);
+		assertEquals("xsappname", configuration.xsuaaAppName1);
+		assertEquals("application", configuration.xsuaaPlan1.toLowerCase());
 
 		/* Index 2 */
 		assertEquals("none", configuration.xsuaaClientId2);
