+++ spring-security/src/test/java/com/sap/cloud/security/spring/config/IdentityServicesPropertySourceFactoryTest.java
@@ -5,7 +5,6 @@
  */
 package com.sap.cloud.security.spring.config;
 
-import com.sap.cloud.security.config.ServiceConstants;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -30,14 +29,14 @@ class IdentityServicesPropertySourceFactoryTest {
 		assertEquals("http://domain.xsuaadomain", configuration.xsuaaUrl);
 		assertEquals("xsuaadomain", configuration.xsuaaDomain);
 		assertEquals("xsappname", configuration.xsuaaAppName);
-		assertEquals(ServiceConstants.Plan.APPLICATION, ServiceConstants.Plan.from(configuration.xsuaaPlan));
+		assertEquals("application", configuration.xsuaaPlan.toLowerCase());
 
 		assertEquals("", configuration.unknown);
 
 		assertEquals("client-id-ias", configuration.identityClientId);
 		assertEquals("client-secret-ias", configuration.identityClientSecret);
 		assertEquals("iasdomain", configuration.identityDomains.get(0));
-		assertEquals(ServiceConstants.Plan.BROKER, ServiceConstants.Plan.from(configuration.iasPlan));
+		assertEquals("broker", configuration.iasPlan.toLowerCase());
 	}
 }
