+++ spring-security/src/test/java/com/sap/cloud/security/spring/config/XsuaaServiceConfigurationLoadingIntegrationTest.java
@@ -6,7 +6,6 @@
 package com.sap.cloud.security.spring.config;
 
 import com.sap.cloud.environment.servicebinding.SapVcapServicesServiceBindingAccessor;
-import com.sap.cloud.security.config.Service;
 import com.sap.cloud.security.config.ServiceBindingEnvironment;
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.BeforeAll;
@@ -18,42 +17,35 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 
 import java.io.IOException;
-import java.util.List;
 
-import static com.sap.cloud.security.spring.config.ConfigurationUtil.assertConfigsAreEqual;
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@SpringBootTest(classes = { MultipleXsuaaConfigurationsFromFile.class })
-class XsuaaServiceConfigurationsLoadingIntegrationTest {
+/**
+ * Tests the integration between XsuaaServiceConfiguration and IdentityServicesPropertySourceFactory.
+ * The XSUAA configuration properties of the XsuaaServiceConfiguration are asserted to be equal to those of the configuration used to populate the Spring properties via IdentityServicesPropertySourceFactory.
+ */
+@SpringBootTest(classes = { SingleXsuaaConfigurationFromFile.class })
+class XsuaaServiceConfigurationLoadingIntegrationTest {
 
 	static ServiceBindingEnvironment env;
 
 	@Autowired
-	XsuaaServiceConfigurations configs;
+	XsuaaServiceConfiguration xsuaaConfig;
 
 	@BeforeAll
 	static void setup() throws IOException {
-		String serviceBindingJson = IOUtils.resourceToString("/fourXsuaaBindingsAndOneIasBinding.json", UTF_8);
+		String serviceBindingJson = IOUtils.resourceToString("/singleXsuaaAndIasBinding.json", UTF_8);
 		env = new ServiceBindingEnvironment(new SapVcapServicesServiceBindingAccessor(any -> serviceBindingJson));
 	}
 
 	@Test
-	void configuresXsuaaServiceConfigurations() {
-		List<XsuaaServiceConfiguration> configList = configs.getConfigurations();
-
-		/* Index 0 backward-compatible behaviour */
-		assertConfigsAreEqual(configList.get(0), env.getXsuaaConfiguration());
-
-		/* Index 1 backward-compatible behaviour */
-		assertConfigsAreEqual(configList.get(1), env.getXsuaaConfigurationForTokenExchange());
-
-		/* Index 2+ */
-		assertEquals(env.getServiceConfigurationsAsList().get(Service.XSUAA).size(), configList.size());
+	void configuresXsuaaServiceConfiguration() {
+		ConfigurationUtil.assertXsuaaConfigsAreEqual(xsuaaConfig, env.getXsuaaConfiguration());
 	}
+
 }
 
 @Configuration
-@PropertySource(factory = IdentityServicesPropertySourceFactory.class, value = { "classpath:fourXsuaaBindingsAndOneIasBinding.json" })
-@EnableConfigurationProperties(XsuaaServiceConfigurations.class)
-class MultipleXsuaaConfigurationsFromFile {}
+@PropertySource(factory = IdentityServicesPropertySourceFactory.class, value = { "classpath:singleXsuaaAndIasBinding.json" })
+@EnableConfigurationProperties(XsuaaServiceConfiguration.class)
+class SingleXsuaaConfigurationFromFile {}
