+++ spring-security/src/test/java/com/sap/cloud/security/spring/config/XsuaaServiceConfigurationsLoadingIntegrationTest.java
@@ -6,6 +6,7 @@
 package com.sap.cloud.security.spring.config;
 
 import com.sap.cloud.environment.servicebinding.SapVcapServicesServiceBindingAccessor;
+import com.sap.cloud.security.config.Service;
 import com.sap.cloud.security.config.ServiceBindingEnvironment;
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.BeforeAll;
@@ -17,31 +18,47 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 
 import java.io.IOException;
+import java.util.List;
 
+import static com.sap.cloud.security.spring.config.ConfigurationUtil.assertXsuaaConfigsAreEqual;
 import static java.nio.charset.StandardCharsets.UTF_8;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@SpringBootTest(classes = { SingleXsuaaConfigurationFromFile.class })
-class XsuaaServiceConfigurationLoadingIntegrationTest {
+/**
+ * Tests the integration between XsuaaServiceConfigurations and IdentityServicesPropertySourceFactory.
+ * The XSUAA configuration properties of the XsuaaServiceConfigurations are asserted to be equal to those of the configuration used to populate the Spring properties via IdentityServicesPropertySourceFactory.
+ * In addition, to assert backward compatibility, the test makes assertions about indices 0 and 1 in the configuration list of XsuaaServiceConfigurations.
+ */
+@SpringBootTest(classes = { MultipleXsuaaConfigurationsFromFile.class })
+class XsuaaServiceConfigurationsLoadingIntegrationTest {
 
 	static ServiceBindingEnvironment env;
 
 	@Autowired
-	XsuaaServiceConfiguration xsuaaConfig;
+	XsuaaServiceConfigurations configs;
 
 	@BeforeAll
 	static void setup() throws IOException {
-		String serviceBindingJson = IOUtils.resourceToString("/singleXsuaaAndIasBinding.json", UTF_8);
+		String serviceBindingJson = IOUtils.resourceToString("/fourXsuaaBindingsAndOneIasBinding.json", UTF_8);
 		env = new ServiceBindingEnvironment(new SapVcapServicesServiceBindingAccessor(any -> serviceBindingJson));
 	}
 
 	@Test
-	void configuresXsuaaServiceConfiguration() {
-		ConfigurationUtil.assertConfigsAreEqual(xsuaaConfig, env.getXsuaaConfiguration());
-	}
+	void configuresXsuaaServiceConfigurations() {
+		List<XsuaaServiceConfiguration> configList = configs.getConfigurations();
 
+		/* Index 0 backward-compatible behaviour */
+		assertXsuaaConfigsAreEqual(configList.get(0), env.getXsuaaConfiguration());
+
+		/* Index 1 backward-compatible behaviour */
+		assertXsuaaConfigsAreEqual(configList.get(1), env.getXsuaaConfigurationForTokenExchange());
+
+		/* Index 2+ */
+		assertEquals(env.getServiceConfigurationsAsList().get(Service.XSUAA).size(), configList.size());
+	}
 }
 
 @Configuration
-@PropertySource(factory = IdentityServicesPropertySourceFactory.class, value = { "classpath:singleXsuaaAndIasBinding.json" })
-@EnableConfigurationProperties(XsuaaServiceConfiguration.class)
-class SingleXsuaaConfigurationFromFile {}
+@PropertySource(factory = IdentityServicesPropertySourceFactory.class, value = { "classpath:fourXsuaaBindingsAndOneIasBinding.json" })
+@EnableConfigurationProperties(XsuaaServiceConfigurations.class)
+class MultipleXsuaaConfigurationsFromFile {}
